Support compiling with -Wdouble-promotion. Add missing #include.

+++ test/test_fixed_trig.cpp
@@ -42,7 +42,7 @@ void TestTrig() {
         double atan2_result = atan2(y, x);
 
         atan2_result += 2 * M_PI;
-        atan2_result = fmod(atan2_result, 2 * M_PI);
+        atan2_result = fmodf(atan2_result, 2 * M_PI);
 
         AssertNearlyEqual(atan2_result, fixed_atan2_result);
       }
